refactor: Move iNetwork functions in Toolkit back to iNetwork Class
