+++ lib/Deprecated.js
@@ -25,6 +25,7 @@
 /* eslint-disable max-classes-per-file */
 
 const deprecate = require('depd');
+const { parseString } = require('xml2js');
 
 const iPgmDeprecate = deprecate('iPgm');
 const iSqlDeprecate = deprecate('iSql');
@@ -474,19 +475,357 @@ class iDataQueue {
 }
 
 class iNetwork {
+  /**
+   * @constructor
+   * @param {Connection} conn
+   */
   constructor(conn) {
-    iNetworkDeprecate('As of v1.0, \'iNetwork Class\' is deprecated. Please use \'Toolkit Class\' instead.');
-    this.toolkit = new Toolkit(conn);
+    iNetworkDeprecate('As of v1.0, class \'iNetwork\' is deprecated and will be removed at a later time.');
+    if (!conn || !typeof conn === 'object') {
+      throw new Error('Expected a Connection object');
+    }
+
+    if (conn.constructor.name === 'Connection') {
+      this.conn = conn;
+      this.reportError = conn.returnError;
+    } else if (conn.constructor.name === 'iConn') {
+      this.conn = conn.connection;
+      this.reportError = this.conn.returnError;
+    } else {
+      throw new Error('Expected a Connection object');
+    }
+
+    /* when returnError is false connection.run((xmloutput))
+     * we create a new Connection object so connection.run((error, xmloutput))
+     * this.reportError is checked within the callback to notify the user or not
+    */
+    if (!this.reportError) {
+      this.conn = new Connection({
+        transport: this.conn.transport,
+        transportOptions: this.conn.transportOptions,
+      });
+    }
+    this.errno = {
+      type: 'ds',
+      io: 'both',
+      len: 'rec2',
+      fields: [
+        {
+          name: 'bytes_provided',
+          type: '10i0',
+          value: 0,
+          setlen: 'rec2',
+        },
+        { name: 'bytes_available', type: '10i0', value: 0 },
+        { name: 'msgid', type: '7A', value: '' },
+        { type: '1A', value: '' },
+      ],
+    };
   }
 
+  /**
+   * @param {function} cb
+   */
   getTCPIPAttr(cb) {
-    iNetworkDeprecate('As of v1.0, \'iNetwork.getTCPIPAttr()\' is deprecated. Please use \'Toolkit.getTCPIPAttr()\' instead.');
-    this.toolkit.getTCPIPAttr(cb);
+    iNetworkDeprecate('As of v1.0, \'iNetwork.getTCPIPAttr()\' is deprecated and will be removed at a later time.');
+    const outBuf = {
+      type: 'ds',
+      io: 'out',
+      len: 'rec1',
+      fields: [
+        { type: '10i0', value: 0 }, // Bytes returned
+        { type: '10i0', value: 0 }, // Bytes available
+        { type: '10i0', value: 0 }, // TCP/IPv4 stack status
+        { type: '10i0', value: 0 }, // How long active
+        { type: '8A', value: '' }, // When last started - date
+        { type: '6A', value: '' }, // When last started - time
+        { type: '8A', value: '' }, // When last ended - date
+        { type: '6A', value: '' }, // When last ended - time
+        { type: '10A', value: '' }, // Who last started - job name
+        { type: '10A', value: '' }, // Who last started - job user name
+        { type: '6A', value: '' }, // Who last started - job number
+        { type: '16h', value: '' }, // Who last started - internal job identifier
+        { type: '10A', value: '' }, // Who last ended - job name
+        { type: '10A', value: '' }, // Who last ended - job user name
+        { type: '6A', value: '' }, // Who last ended - job number
+        { type: '16h', value: '' }, // Who last ended - internal job identifier
+        { type: '10i0', value: 0 }, // Offset to additional information
+        { type: '10i0', value: 0 }, // Length of additional information
+        { type: '10i0', value: 0 }, // Limited mode
+        { type: '10i0', value: 0 }, // Offset to list of Internet addresses
+        { type: '10i0', value: 0 }, // Number of Internet addresses
+        { type: '10i0', value: 0 }, // Entry length for list of Internet addresses
+        { type: '10i0', value: 0 }, // DNS protocol
+        { type: '10i0', value: 0 }, // Retries
+        { type: '10i0', value: 0 }, // Time interval
+        { type: '10i0', value: 0 }, // Search order
+        { type: '10i0', value: 0 }, // Initial domain name server
+        { type: '10i0', value: 0 }, // DNS listening port
+        { type: '64A', value: '' }, // Host name
+        { type: '255A', value: '' }, // Domain name
+        { type: '1A', value: '' }, // Reserved
+        { type: '256A', value: '' }, // ]Domain search list
+      ],
+    };
+
+    const pgm = new ProgramCall('QTOCNETSTS', { lib: 'QSYS', func: 'QtocRtvTCPA' });
+    pgm.addParam(outBuf);
+    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
+    pgm.addParam({ type: '8A', value: 'TCPA0300' });
+    pgm.addParam(this.errno);
+    this.conn.add(pgm.toXML());
+
+    let rtValue; // The returned value.
+
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          const { data } = result.myscript.pgm[0].parm[0].ds[0];
+          rtValue = {
+            'TCP/IPv4_stack_status': data[2]._,
+            How_long_active: data[3]._,
+            'When_last_started_-_date': data[4]._,
+            'When_last_started_-_time': data[5]._,
+            'When_last_ended_-_date': data[6]._,
+            'When_last_ended_-_time': data[7]._,
+            'Who_last_started_-_job_name': data[8]._,
+            'Who_last_started_-_job_user_name': data[9]._,
+            'Who_last_started_-_job_number': data[10]._,
+            'Who_last_started_-_internal_job_identifier': data[11]._,
+            'Who_last_ended_-_job_name': data[12]._,
+            'Who_last_ended_-_job_user_name': data[13]._,
+            'Who_last_ended_-_job_number': data[14]._,
+            'Who_last_ended_-_internal_job_identifier': data[14]._,
+            Offset_to_additional_information: data[16]._,
+            Length_of_additional_information: data[17]._,
+            Limited_mode: data[18]._,
+            Offset_to_list_of_Internet_addresses: data[19]._,
+            Number_of_Internet_addresses: data[20]._,
+            Entry_length_for_list_of_Internet_addresses: data[21]._,
+            DNS_protocol: data[22]._,
+            Retries: data[23]._,
+            Time_interval: data[24]._,
+            Search_order: data[25]._,
+            Initial_domain_name_server: data[26]._,
+            DNS_listening_port: data[27]._,
+            Host_name: data[28]._,
+            Domain_name: data[29]._,
+            Reserved: data[30]._,
+            Domain_search_list: data[31]._,
+          };
+        } else { rtValue = str; }
+
+        if (this.reportError) {
+          cb(null, rtValue);
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 
+  /**
+   * @param {string} ipaddr
+   * @param {function} cb
+   */
   getNetInterfaceData(ipaddr, cb) {
-    iNetworkDeprecate('As of v1.0, \'iNetwork.getNetInterfaceData()\' is deprecated. Please use \'Toolkit.getNetInterfaceData()\' instead.');
-    this.toolkit.getNetInterfaceData(ipaddr, cb);
+    iNetworkDeprecate('As of v1.0, \'iNetwork.getNetInterfaceData()\' is deprecated and will be removed at a later time.');
+    let address = ipaddr;
+
+    const IPToNum = (ip) => {
+      const ipArray = ip.split('.');
+      let num = 0;
+
+      /* eslint-disable no-bitwise */
+      num += parseInt(ipArray[0], 10) << 24;
+      num += parseInt(ipArray[1], 10) << 16;
+      num += parseInt(ipArray[2], 10) << 8;
+      num += parseInt(ipArray[3], 10);
+      num >>>= 0; // zero-fill right-shift, returns non-negative
+
+      return num;
+    };
+    const outBuf = {
+      type: 'ds',
+      io: 'out',
+      len: 'rec1',
+      fields: [
+        { type: '10i0', value: 0 }, // Bytes returned
+        { type: '10i0', value: 0 }, // Bytes available
+        { type: '15A', value: '' }, // Internet address
+        { type: '1A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Internet address binary
+        { type: '15A', value: '' }, // Network address
+        { type: '1A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Network address binary
+        { type: '10A', value: '' }, // Line description
+        { type: '2A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Interface status
+        { type: '10i0', value: 0 }, // Interface type of service
+        { type: '10i0', value: 0 }, // Interface MTU
+        { type: '10i0', value: 0 }, // Interface line type
+        { type: '15A', value: '' }, // Host address
+        { type: '1A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Host address binary
+        { type: '15A', value: '' }, // Interface subnet mask
+        { type: '1A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Interface subnet mask binary
+        { type: '15A', value: '' }, // Directed broadcast address
+        { type: '1A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Directed broadcast address binary
+        { type: '8A', value: '' }, // Change date
+        { type: '6A', value: '' }, // Change time
+        { type: '15A', value: '' }, // Associated local interface
+        { type: '3A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Associated local interface binary
+        { type: '10i0', value: 0 }, // Change status
+        { type: '10i0', value: 0 }, // Packet rules
+        { type: '10i0', value: 0 }, // Automatic start
+        { type: '10i0', value: 0 }, // TRLAN bit sequencing
+        { type: '10i0', value: 0 }, // Interface type
+        { type: '10i0', value: 0 }, // Proxy ARP allowed
+        { type: '10i0', value: 0 }, // Proxy ARP enabled
+        { type: '10i0', value: 0 }, // Configured MTU
+        { type: '24A', value: '' }, // Network name
+        { type: '24A', value: '' }, // Interface name
+        { type: '50A', value: '' }, // Alias name
+        { type: '2A', value: '' }, // Reserved
+        { type: '50A', value: '' }, // Interface description
+        { type: '2A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Offset to preferred interface list
+        { type: '10i0', value: 0 }, // Number of entries in preferred interface list
+        { type: '10i0', value: 0 }, // Length of one preferred interface list entry
+        { type: '10i0', value: 0 }, // DHCP created
+        { type: '10i0', value: 0 }, // DHCP dynamic DNS updates
+        { type: '20i0', value: 0 }, // DHCP lease expiration
+        { type: '8A', value: '' }, // DHCP lease expiration - date
+        { type: '6A', value: '' }, // DHCP lease expiration - time
+        { type: '20i0', value: 0 }, // DHCP lease obtained
+        { type: '8A', value: '' }, // DHCP lease obtained - date
+        { type: '6A', value: '' }, // DHCP lease obtained - time
+        { type: '10i0', value: 0 }, // Use DHCP unique identifier
+        { type: '15A', value: '' }, // DHCP server IP address
+        { type: '1A', value: '' }, // Reserved
+        { type: '15h', value: '' }, // Preferred interface Internet address
+        { type: '1A', value: '' }, // Reserved
+        { type: '10i0', value: 0 }, // Preferred interface Internet address binary
+        { type: '1A', value: '' }, // Reserved
+      ],
+    };
+    if (ipaddr === 'localhost') { address = '127.0.0.1'; }
+    const request = {
+      type: 'ds',
+      fields: [
+        { type: '10i0', value: '1' },
+        { type: '10i0', value: IPToNum(address) },
+      ],
+    };
+    const pgm = new ProgramCall('QTOCNETSTS', { lib: 'QSYS', func: 'QtocRtvNetIfcDta' });
+    pgm.addParam(outBuf);
+    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
+    pgm.addParam({ type: '8A', value: 'NIFD0100' });
+    pgm.addParam(request);
+    pgm.addParam(this.errno);
+    this.conn.add(pgm.toXML());
+
+    let rtValue; // The returned value.
+
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          const { data } = result.myscript.pgm[0].parm[0].ds[0];
+
+          rtValue = {
+            Internet_address: data[2]._,
+            Internet_address_binary: data[4]._,
+            Network_address: data[5]._,
+            Network_address_binary: data[7]._,
+            Line_description: data[8]._,
+            Interface_status: data[10]._,
+            Interface_type_of_service: data[11]._,
+            Interface_MTU: data[12]._,
+            Interface_line_type: data[13]._,
+            Host_address: data[14]._,
+            Host_address_binary: data[16]._,
+            Interface_subnet_mask: data[17]._,
+            Interface_subnet_mask_binary: data[19]._,
+            Directed_broadcast_address: data[20]._,
+            Directed_broadcast_address_binary: data[22]._,
+            Change_date: data[23]._,
+            Change_time: data[24]._,
+            Associated_local_interface: data[25]._,
+            Associated_local_interface_binary: data[27]._,
+            Change_status: data[28]._,
+            Packet_rules: data[29]._,
+            Automatic_start: data[30]._,
+            TRLAN_bit_sequencing: data[31]._,
+            Interface_type: data[32]._,
+            Proxy_ARP_allowed: data[33]._,
+            Proxy_ARP_enabled: data[34]._,
+            Configured_MTU: data[35]._,
+            Network_name: data[36]._,
+            Interface_name: data[37]._,
+            Alias_name: data[38]._,
+            Interface_description: data[40]._,
+            Offset_to_preferred_interface_list: data[42]._,
+            Number_of_entries_in_preferred_interface_list: data[43]._,
+            Length_of_one_preferred_interface_list_entry: data[44]._,
+            DHCP_created: data[45]._,
+            DHCP_dynamic_DNS_updates: data[46]._,
+            DHCP_lease_expiration: data[47]._,
+            'DHCP_lease_expiration_-_date': data[48]._,
+            'DHCP_lease_expiration_-_time': data[49]._,
+            DHCP_lease_obtained: data[50]._,
+            'DHCP_lease_obtained_-_date': data[51]._,
+            'DHCP_lease_obtained_-_time': data[52]._,
+            Use_DHCP_unique_identifier: data[53]._,
+            DHCP_server_IP_address: data[54]._,
+            Preferred_interface_Internet_address: data[56]._,
+            Preferred_interface_Internet_address_binary: data[58]._,
+          };
+        } else { rtValue = str; }
+
+        if (this.reportError) {
+          cb(null, rtValue);
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 }
